feat: Move Events to Shared project
