+++ CleanArchitecture.Domain.Tests/CommandHandler/Tenant/DeleteTenant/DeleteTenantCommandHandlerTests.cs
@@ -1,7 +1,7 @@
 using System;
 using CleanArchitecture.Domain.Commands.Tenants.DeleteTenant;
 using CleanArchitecture.Domain.Errors;
-using CleanArchitecture.Domain.Events.Tenant;
+using CleanArchitecture.Shared.Events.Tenant;
 using Xunit;
 
 namespace CleanArchitecture.Domain.Tests.CommandHandler.Tenant.DeleteTenant;
