+++ CleanArchitecture.Domain.Tests/CommandHandler/User/ChangePassword/ChangePasswordCommandHandlerTests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using CleanArchitecture.Domain.Commands.Users.ChangePassword;
 using CleanArchitecture.Domain.Errors;
-using CleanArchitecture.Domain.Events.User;
+using CleanArchitecture.Shared.Events.User;
 using Xunit;
 
 namespace CleanArchitecture.Domain.Tests.CommandHandler.User.ChangePassword;
