+++ CleanArchitecture.Domain.Tests/CommandHandlerFixtureBase.cs
@@ -1,9 +1,9 @@
 using System;
 using System.Linq.Expressions;
-using CleanArchitecture.Domain.DomainEvents;
 using CleanArchitecture.Domain.Enums;
 using CleanArchitecture.Domain.Interfaces;
 using CleanArchitecture.Domain.Notifications;
+using CleanArchitecture.Shared.Events;
 using NSubstitute;
 
 namespace CleanArchitecture.Domain.Tests;
