+++ CleanArchitecture.Domain/CleanArchitecture.Domain.csproj
@@ -18,4 +18,8 @@
     <ItemGroup>
         <FrameworkReference Include="Microsoft.AspNetCore.App" />
     </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\CleanArchitecture.Shared\CleanArchitecture.Shared.csproj" />
+    </ItemGroup>
 </Project>
