+++ CleanArchitecture.Domain/DomainEvents/StoredDomainEvent.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using CleanArchitecture.Shared.Events;
 
 namespace CleanArchitecture.Domain.DomainEvents;
