+++ CleanArchitecture.Domain/Interfaces/IMediatorHandler.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 using CleanArchitecture.Domain.Commands;
-using CleanArchitecture.Domain.DomainEvents;
+using CleanArchitecture.Shared.Events;
 using MediatR;
 
 namespace CleanArchitecture.Domain.Interfaces;
