+++ CleanArchitecture.Infrastructure.Tests/InMemoryBusTests.cs
@@ -3,8 +3,8 @@ using System.Threading.Tasks;
 using CleanArchitecture.Domain.Commands.Users.DeleteUser;
 using CleanArchitecture.Domain.DomainEvents;
 using CleanArchitecture.Domain.EventHandler.Fanout;
-using CleanArchitecture.Domain.Events.User;
 using CleanArchitecture.Domain.Notifications;
+using CleanArchitecture.Shared.Events.User;
 using MediatR;
 using NSubstitute;
 using Xunit;
