+++ CleanArchitecture.Infrastructure/CleanArchitecture.Infrastructure.csproj
@@ -7,6 +7,7 @@
 
     <ItemGroup>
         <ProjectReference Include="..\CleanArchitecture.Domain\CleanArchitecture.Domain.csproj"/>
+        <ProjectReference Include="..\CleanArchitecture.Shared\CleanArchitecture.Shared.csproj" />
     </ItemGroup>
 
     <ItemGroup>
