+++ CleanArchitecture.Shared/CleanArchitecture.Shared.csproj
@@ -6,4 +6,8 @@
     </PropertyGroup>
 
 
+    <ItemGroup>
+      <PackageReference Include="MediatR" Version="12.1.1" />
+    </ItemGroup>
+
 </Project>
