+++ CleanArchitecture.Shared/Events/DomainEvent.cs
@@ -1,9 +1,8 @@
 using System;
 using MediatR;
 
-namespace CleanArchitecture.Domain.DomainEvents;
+namespace CleanArchitecture.Shared.Events;
 
-// Todo: Move this and all events to shared
 public abstract class DomainEvent : Message, INotification
 {
     public DateTime Timestamp { get; private set; }
